Move migrations inside of the Postgres package
